refactor(test-fixtures): consolidate per-fixture example cdylibs into multi-actor bundles
Groups the probe, inline-child, and DAG fixtures into three multi-actor
bundles via ADR-0096 `export!(…)`, reducing the fixture-stanza wall from
16 `[[example]]` entries to 13 and cutting three wasm compilation units:

- `probe` bundle: folds `probe_with_config` (export
  `"test_fixtures_probe_with_config"`). Consumers pass `export: Some(…)`
  on the load call; the entry `Probe` is unchanged for its ~13 callers.
- `inline_child` bundle: folds `inline_child_stateful` (export
  `"test.inline.stateful_parent"`) and `inline_child_despawn` (export
  `"test.inline.despawn_parent"`). The `stateful_replace{,_typed,_reshaped}`
  trio stays separate — the typed→reshaped cross-module replace test
  requires two distinct wasm binaries.
- `mat4_source` bundle: folds `vec4_observer` (export `"vec4_observer"`).

Consumer load-site updates: `typed_config.rs`, `fleetbench_mail.rs`
(probe bundle); `test_bench_scenario.rs` inline-stateful + despawn loads
(inline_child bundle); `fleetbench_transform_dag.rs` (mat4_source bundle).
Two new FleetBench helpers — `load_full_export` and
`load_with_config_export` — provide the export-bearing equivalents of
the existing `load_full` / `load_with_config` so consumers don't need
to drop to `load_by_selector`.

Closes #1994
