+++ crates/aether-test-fixtures/examples/probe.rs
@@ -1,3 +1,9 @@
+//! `probe` bundle — the entry `Probe` fixture plus the ADR-0090 c1
+//! `ProbeWithConfig` typed-config fixture, exported together via
+//! `export!(Probe, ProbeWithConfig)` (ADR-0096, issue 1994).
+//!
+//! # `Probe`
+//!
 //! Test-fixture component for substrate-feature scenarios. Not a
 //! demo, not exemplary — its only job is to expose substrate /
 //! test-bench primitives (input subscription, drop, replace, capture)
@@ -29,18 +35,39 @@
 //! actor source is unchanged; the shared `TickObserved` / `SetRender`
 //! kinds moved to the sibling lib so integration tests can import
 //! them without reaching into a cdylib.
+//!
+//! # `ProbeWithConfig`
+//!
+//! ADR-0090 c1 typed-config fixture. Exercises the
+//! `FfiActor::Config = ProbeConfig` path end-to-end: the host places
+//! postcard-encoded `ProbeConfig` bytes in a delivery region (ADR-0095) during
+//! `Component::instantiate`; the guest's `init_with_config_p32` shim decodes
+//! them via `<ProbeConfig as Kind>::decode_from_bytes` and threads
+//! the typed struct into `Probe::init(config, ctx)`.
+//!
+//! The fixture stashes `(seed, label)` at boot and replies with a
+//! `ConfigEcho` on every `ConfigQuery` mail so a test can assert the
+//! config round-tripped intact. No tick / render behaviour — the
+//! sibling `Probe` covers that.
+//!
+//! Consumers load it from the `probe` bundle stem with
+//! `export: Some("test_fixtures_probe_with_config")` (ADR-0096).
 
 // `on_key` only re-broadcasts the inbound payload, so it doesn't touch
 // `self`; it keeps `&mut self` to match the `#[handler]` dispatch ABI.
+// `ProbeWithConfig::on_config_query` takes `&mut self` for the same reason.
 #![allow(clippy::unused_self)]
 
-use aether_actor::{BootError, FfiActor, FfiCtx, FfiInitCtx, MailSender, actor};
+use aether_actor::{
+    BootError, FfiActor, FfiCtx, FfiInitCtx, MailSender, Manual, OutboundReply, actor,
+};
 use aether_capabilities::input::InputMailboxExt;
 use aether_capabilities::lifecycle::LifecycleMailboxExt;
 use aether_capabilities::{InputCapability, LifecycleCapability, RenderCapability};
 use aether_kinds::{DrawTriangle, Key, Tick, Vertex};
 use aether_test_fixtures::{
-    KeyObserved, SetRender, TEST_BENCH_OBSERVER_MAILBOX_NAME, TickObserved,
+    ConfigEcho, ConfigQuery, KeyObserved, ProbeConfig, SetRender, TEST_BENCH_OBSERVER_MAILBOX_NAME,
+    TickObserved,
 };
 
 pub struct Probe {
@@ -143,4 +170,40 @@ impl FfiActor for Probe {
     }
 }
 
-aether_actor::export!(Probe);
+/// ADR-0090 c1 typed-config fixture. Exercises the
+/// `FfiActor::Config = ProbeConfig` path end-to-end.
+///
+/// Consumers load this actor from the `probe` bundle with
+/// `export: Some("test_fixtures_probe_with_config")`.
+pub struct ProbeWithConfig {
+    seed: u32,
+    label: String,
+}
+
+#[actor]
+impl FfiActor for ProbeWithConfig {
+    type Config = ProbeConfig;
+    const NAMESPACE: &'static str = "test_fixtures_probe_with_config";
+
+    fn init(config: ProbeConfig, _ctx: &mut FfiInitCtx<'_>) -> Result<Self, BootError> {
+        Ok(ProbeWithConfig {
+            seed: config.seed,
+            label: config.label,
+        })
+    }
+
+    /// Reply with a `ConfigEcho` describing the cached config. Lets
+    /// the integration test observe what the typed `init` actually
+    /// received without scraping logs or readback.
+    #[handler::manual]
+    fn on_config_query(&mut self, ctx: &mut FfiCtx<'_, Manual>, _query: ConfigQuery) {
+        if ctx.reply_target().is_some() {
+            ctx.reply(&ConfigEcho {
+                seed: self.seed,
+                label: self.label.clone(),
+            });
+        }
+    }
+}
+
+aether_actor::export!(Probe, ProbeWithConfig);
